+++ src/components/molecules/ChallengeEditCard/index.tsx
@@ -1,17 +1,19 @@
 import { ChallengeEdit, getChallengeContent } from 'apis';
+import {
+  CharacterBlack,
+  CharacterColor1,
+  CharacterColor2,
+  CharacterColor3,
+  MoreClickedIcon,
+} from 'assets/images/index';
 import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import { userStatusState } from 'stores/user';
-import Styled from 'styled-components';
-import CharacterBlack from '../../../assets/images/character_black.svg';
-import CharacterColor1 from '../../../assets/images/character_color1.svg';
-import CharacterColor2 from '../../../assets/images/character_color2.svg';
-import CharacterColor3 from '../../../assets/images/character_color3.svg';
-import MoreIcon from '../../../assets/images/moreIcon.svg';
 import { interestList } from '../../../resources/string';
 import Button from '../../atoms/Button';
 import Modal from '../../atoms/Modal/index';
+import SWriteCard from './style';
 
 export interface IProps {
   description?: string;
@@ -76,24 +78,24 @@ function EditCard({ id }: IEditCard) {
     byte3: 0,
   });
 
-  const [form, setForm] = useState({
+  const [textForm, setTextForm] = useState({
     description1: '',
     description2: '',
     description3: '',
   });
 
-  const { description1, description2, description3 } = form;
-  const onChange2 = (target: any) => {
+  const { description1, description2, description3 } = textForm;
+  const textOnChange = (target: any) => {
     const { name, value } = target;
-    setForm({
-      ...form,
+    setTextForm({
+      ...textForm,
       [name]: value,
     });
   };
 
   const handleSubmit = async (e: any) => {
     e.preventDefault();
-    setForm({
+    setTextForm({
       description1: '',
       description2: '',
       description3: '',
@@ -124,7 +126,7 @@ function EditCard({ id }: IEditCard) {
     setCountProgressBar(count);
   };
 
-  const onChange = (e: any) => {
+  const totalOnChange = (e: any) => {
     const text_val = e.target.value; //입력한 문자
     const target_byte_name = 'byte' + e.target.name.split('description')[1];
     let str = '';
@@ -135,7 +137,7 @@ function EditCard({ id }: IEditCard) {
       e.target.value = str;
       totalByte = maxByte;
     }
-    onChange2(e.target);
+    textOnChange(e.target);
     setByte({
       ...byte,
       [target_byte_name]: totalByte,
@@ -156,7 +158,7 @@ function EditCard({ id }: IEditCard) {
     if (userStatusData) {
       const data = await getChallengeContent(id, userStatusData.token);
       if (data) {
-        setForm({ description1: data.good, description2: data.bad, description3: data.learn });
+        setTextForm({ description1: data.good, description2: data.bad, description3: data.learn });
         setSelectedInterest(data.interest);
         setByte({
           byte1: stringtoByte(data.good),
@@ -170,7 +172,6 @@ function EditCard({ id }: IEditCard) {
 
   const [isOpenTag, setIsOpenTag] = useState(false);
   const [isClickTag, setIsClickTag] = useState(false);
-  const [isInterestModalOpen, setIsInterestModalOpen] = useState(false);
 
   const [selectedInterest, setSelectedInterest] = useState<string[]>([]);
   const modalInterestHandler = (interest: string) => {
@@ -228,8 +229,8 @@ function EditCard({ id }: IEditCard) {
 
         {countProgressBar === 1 ? (
           <div className="character">
-            <div className="character__color1">
-              <img className="character__1" src={CharacterColor1} alt="" />
+            <div className="character__color-step1">
+              <img className="character__detail-step1" src={CharacterColor1} alt="" />
               <div className="character__message">오호라! 오늘 이런 일이 있었군요!</div>
             </div>
             <div className="bar">
@@ -241,8 +242,8 @@ function EditCard({ id }: IEditCard) {
         ) : null}
         {countProgressBar === 2 ? (
           <div className="character">
-            <div className="character__color2">
-              <img className="character__2" src={CharacterColor2} alt="" />
+            <div className="character__color-step2">
+              <img className="character__detail-step2" src={CharacterColor2} alt="" />
               <div className="character__message">우와! 내일의 당신은 더 행복할거에요 :)</div>
             </div>
             <div className="bar">
@@ -254,13 +255,13 @@ function EditCard({ id }: IEditCard) {
         ) : null}
         {countProgressBar === 3 ? (
           <div className="character">
-            <div className="character__color3">
-              <img className="character__3" src={CharacterColor3} alt="" />
+            <div className="character__color-step3">
+              <img className="character__detail-step3" src={CharacterColor3} alt="" />
               <div className="character__message">더 성장한 내일의 나를 위해!</div>
             </div>
             <div className="bar">
               <span className="progressbar">
-                <span className="gauge__one">.</span>
+                <span className="gauge__whole">.</span>
               </span>
             </div>
           </div>
@@ -276,7 +277,7 @@ function EditCard({ id }: IEditCard) {
             name="description1"
             value={description1}
             placeholder="오늘의 잘한 점을 적어보세요."
-            onChange={onChange}
+            onChange={totalOnChange}
           ></textarea>
         </div>
         <div className="challenge-card">
@@ -289,7 +290,7 @@ function EditCard({ id }: IEditCard) {
             name="description2"
             value={description2}
             placeholder="오늘의 못한 점을 적어보세요."
-            onChange={onChange}
+            onChange={totalOnChange}
           ></textarea>
         </div>
         <div className="challenge-card">
@@ -302,7 +303,7 @@ function EditCard({ id }: IEditCard) {
             name="description3"
             value={description3}
             placeholder="배운 것과 실천할 것을 적어보세요."
-            onChange={onChange}
+            onChange={totalOnChange}
           ></textarea>
         </div>
 
@@ -320,11 +321,11 @@ function EditCard({ id }: IEditCard) {
                 setIsOpenTag(!isOpenTag);
               }}
             >
-              <img className="tag__moreIcon" src={MoreIcon} alt=""></img>
+              <img className="tag__moreIcon" src={MoreClickedIcon} alt=""></img>
             </Button>
-            <div className="tag__group1">
+            <div className="tag__group">
               {isOpenTag === true ? (
-                <button className="tag__group2">
+                <button className="tag__group-detail">
                   {isOpenTag === true ? (
                     interestList.map((interest, id) => {
                       return setInterestButton(id, interest);
@@ -361,195 +362,5 @@ function EditCard({ id }: IEditCard) {
     </>
   );
 }
-
-const SWriteCard = Styled.div`
-
-    .header{
-      padding-top:100px;
-      text-align:center;
-      line-height: 1.22;
-      letter-spacing: -0.5px;
-      font-family: HomepageBaukasten;
-      font-size: 46px;
-      font-weight: bold;
-}
-    .challenge-card{
-        margin: 0 auto;
-        padding-top:60px;
-        width:844px;
-
-        &__title{
-            line-height:38px;
-            letter-spacing: -0.5px;
-            color: #3D3D3D;
-            font-size:22px;
-            font-weight: bold;
-        }
-        &__restriction{
-            padding-bottom: 30px;
-            text-align: right;
-            line-height: 18px;
-            color: #B1B1B1;
-            font-size: 14px;
-            font-weight: bold;
-            
-        }
-    }
-    .textarea{
-        box-sizing: border-box;
-        align-items: center;
-        border: 1px solid #DFDFDF;
-        padding: 40px 60px 40px 60px;
-        width: 844px;
-        height:369px;
-        resize:none;
-        font-size: 18px;
-    }
-
-    .button{
-        align-items: center;
-        margin: 0 auto;
-        width:844px;
-    }
-    
-    .write__button-color{
-        margin: 0 auto;
-        margin-top:60px;
-        border:none;
-        border-radius: 4px;
-        background: linear-gradient(91.91deg, #36C8F5 7.34%, #13E2DD 90.35%);
-        width:844px;
-        height: 60px;
-        text-align: center;
-        color: #FFFFFF;
-    }
-    .write__button-black{
-        margin: 0 auto;
-        margin-top:60px;
-        border:none;
-        border-radius: 4px;
-        background: #DFDFDF;
-        width:844px;
-        height: 60px;
-        text-align: center;
-        color: #FFFFFF;
-    }
-    .character{
-        position:sticky;
-        top:20px;
-        margin: 0 auto;
-        background-color:white;
-        padding-top:60px;
-        width:844px;
-
-        &__1{
-            padding-left:60px;
-        }
-        &__2{
-            padding-left:50px;
-        }
-        &__black{
-            padding-top:29px;
-            padding-right:50px;
-        }
-        &__color1{
-            padding-top:29px;
-            padding-left:200px;
-        }
-        &__color2{
-            padding-left:450px;
-        }
-        &__color3{
-            text-align:end;
-        }
-
-        &__message{
-            padding-top:12.32px;
-            padding-bottom: 12px;
-            line-height: 1.43;
-            letter-spacing: -0.5px;
-            font-size: 14px;
-            font-weight: bold;
-        }
-    }   
-    .progressbar{
-        display: inline-block;
-        background-color: #d6d3d3;
-        width: 844px;
-        height: 5px;
-    }
-    .gauge__initial{
-        display: inline-block;
-        background-color: #3abff7;
-        width: 1%;
-        height: 5px;
-    }
-
-    .gauge__quarter{
-            display: inline-block;
-            background-color: #3abff7;
-            width: 34%;
-            height: 5px;
-        }
-    .gauge__half{
-            display: inline-block;
-            background-color: #3abff7;
-            width: 67%;
-            height: 5px;
-        }
-    .gauge__one{
-            display: inline-block;
-            background-color: #3abff7;
-            width: 100%;
-            height: 5px;
-        }
-
-    .tag{
-      align-items:center;
-      margin:0 auto;
-      padding-top:60px;
-      width:844px;
-      text-align: left;
-      line-height: 1.33;
-      letter-spacing: -0.5px;
-      color: var(--colors-grayscale-6-f);
-      font-family: AppleSDGothicNeo;
-      font-size: 24px;
-      font-weight: bold;
-
-      &__moreIcon{
-        padding-bottom:5px;
-        vertical-align : middle;
-      }
-  
-      &__interest{
-          margin : 12px 5px;
-          border: solid 1px #8b8b8b;
-          border-radius: 60px;
-          background-color:#FFFFFF;
-          padding : 12px 30px;
-          height : 48px;
-          line-height: 1.33;
-          letter-spacing: -0.5px;
-          color : #8b8b8b;
-          font-size: 18px;
-          font-weight: bold;
-        }
-
-        .tag__group1{
-          padding-top:20px;
-
-        }
-        .tag__group2{
-          box-sizing: border-box;
-          border: 1px solid rgba(223, 223, 223, 0.5);
-          border-radius: 16px;
-          box-shadow: 0px 0px 15px rgba(23, 22, 91, 0.08);
-          background: #FFFFFF;
-          padding: 20px 30px 30px 30px
-          
-        }
-
-}`;
 
 export default EditCard;
